🤖 refactor(site/src/pages): nest template creation pages
