+++ site/src/modules/templates/VariableInput/VariableInput.tsx
