+++ site/src/pages/TemplatesPage/CreateTemplatePage/DuplicateTemplateView.tsx
