+++ site/src/pages/TemplatesPage/CreateTemplatePage/TemplateBuilderPage/SelectionSummary.tsx
