+++ site/src/pages/TemplatesPage/CreateTemplatePage/TemplateBuilderPage/TemplateAlternatives.stories.tsx
@@ -2,7 +2,8 @@ import type { Meta, StoryObj } from "@storybook/react-vite";
 import { TemplateAlternatives } from "./TemplateAlternatives";
 
 const meta: Meta<typeof TemplateAlternatives> = {
-	title: "pages/TemplateBuilder/TemplateAlternatives",
+	title:
+		"pages/TemplatesPage/CreateTemplatePage/TemplateBuilderPage/TemplateAlternatives",
 	component: TemplateAlternatives,
 };
