+++ site/src/pages/TemplatesPage/CreateTemplatePage/TemplateBuilderPage/TemplateAlternatives.tsx
