+++ site/src/pages/TemplatesPage/CreateTemplatePage/TemplateBuilderPage/TemplateBuilderAvatarData.tsx
