+++ site/src/pages/TemplatesPage/CreateTemplatePage/TemplateBuilderPage/TemplateBuilderPageView.tsx
