+++ site/src/pages/TemplatesPage/CreateTemplatePage/TemplateBuilderPage/TemplateCard.tsx
