+++ site/src/pages/TemplatesPage/CreateTemplatePage/TemplateBuilderPage/defaultPlaceholder.test.ts
