+++ site/src/pages/TemplatesPage/CreateTemplatePage/TemplateBuilderPage/defaultPlaceholder.ts
