+++ site/src/pages/TemplatesPage/CreateTemplatePage/TemplateBuilderPage/hooks/useFuzzySearch.ts
